+++ R/plot.R
@@ -10,7 +10,7 @@
 #'
 #' @examples
 #' m <- create_model(AirPassengers, lags = 1:12, method = "rf")
-#' f <- FORECAST(m, h = 12)
+#' f <- forecast(m, h = 12)
 #' library(ggplot2)
 #' autoplot(f)
 #' @export