+++ R/first_differences.R
@@ -1,37 +1,3 @@
-#' Specifying the transformation for dealing with trended series
-#'
-#' This function is used to specify the preprocessing associated with the trend
-#' of a time series.
-#'
-#' @param type A character indicating the type of preprocessing applied to the
-#'   time series. Possible values are: `"none"`, `"additive"`, `"multiplicative"`
-#'   and `"differences"`.
-#' @param n An integer specifying the order of first differences to be applied.
-#'   If the default (-1) is used, the order of first differences needed by the
-#'   time series will be estimated by the [forecast::ndiffs()] function. This
-#'   parameter is only meant when the `type` parameter is `"differences"`.
-#' @param transform_features A logical value indicating whether the training
-#'   features are also transformed with the additive or multiplicative
-#'   transformation.
-#'
-#' @return A list with the selected options
-#' @export
-#'
-#' @examples
-#' trend("none")            # no preprocessing
-#' trend("additive")        # additive preprocessing
-#' trend("differences", 1)  # order 1 first differences
-#' trend("differences", -1) # order of first differences automatically estimated
-trend <- function(type = c("additive", "multiplicative", "differences", "none"), 
-                  n = -1, transform_features = TRUE) {
-  type <- match.arg(type)
-  if (! (is.numeric(n) && length(n) == 1 && n >= -1 && floor(n) == n))
-    stop("n parameter should be an integer scalar value >= -1")
-  if (! (is.logical(transform_features) && length(transform_features) == 1))
-    stop("transform_features parameter should be a logical value")
-  structure(list(type = type, n = n, transform_features = transform_features), class = "trend")
-}
-
 # Differences preprocessing
 # timeS is the time series
 # differences is the number of differences, if -1, the number of differences is